+++ src/src/Globals/Plugins.cpp
@@ -30,7 +30,7 @@ std::map<pluginID_t, deviceIndex_t> Plugin_id_to_DeviceIndex;
 std::vector<pluginID_t>    DeviceIndex_to_Plugin_id;
 std::vector<deviceIndex_t> DeviceIndex_sorted;
 
-float customFloatVar[CUSTOM_VARS_MAX];
+double customFloatVar[CUSTOM_VARS_MAX];
 
 float UserVar[VARS_PER_TASK * TASKS_MAX];
