+++ src/src/Globals/Plugins.h
@@ -51,7 +51,7 @@
 * let,1,10
 * if %v1%=10 do ...
 \*********************************************************************************************/
-extern float customFloatVar[CUSTOM_VARS_MAX];
+extern double customFloatVar[CUSTOM_VARS_MAX];
 extern float UserVar[VARS_PER_TASK * TASKS_MAX];
 
 extern int deviceCount;
